Added "IsAccessibilityObjectCreated" property
The "IsAccessibilityObjectCreated" property is needed to get data about whether an accessibility object has been created. Currently, we only have access to the accessibility object through the "AccessibilityObject" property, which forces the accessibility object to be created when we try to access it.

The "IsAccessibilityObjectCreated" property will help avoid using memory by unnecessary accessibility objects and prevent execution of unnecessary code when external accessibility tools are not active